+++ src/compiler/transformers/collections/parse-collection-manifest.ts
@@ -18,6 +18,7 @@ export const parseCollectionManifest = (
   const collection: d.CollectionCompilerMeta = {
     collectionName: collectionName,
     moduleId: collectionName,
+    moduleDir: collectionDir,
     moduleFiles: [],
     dependencies: parseCollectionDependencies(collectionManifest),
     compiler: {