+++ src/declarations/stencil-private.ts
@@ -26,6 +26,7 @@ import type {
   PrerenderConfig,
   StyleDoc,
   TaskCommand,
+  ValidatedConfig,
 } from './stencil-public-compiler';
 import type { ComponentInterface, ListenTargetOptions, VNode } from './stencil-public-runtime';
 
@@ -220,7 +221,7 @@ export interface BuildCtx {
   buildStats?: result.Result<CompilerBuildStats, { diagnostics: Diagnostic[] }>;
   buildMessages: string[];
   bundleBuildCount: number;
-  collections: Collection[];
+  collections: CollectionCompilerMeta[];
   compilerCtx: CompilerCtx;
   esmBrowserComponentBundle: ReadonlyArray<BundleModule>;
   esmComponentBundle: ReadonlyArray<BundleModule>;
@@ -229,7 +230,7 @@ export interface BuildCtx {
   commonJsComponentBundle: ReadonlyArray<BundleModule>;
   components: ComponentCompilerMeta[];
   componentGraph: Map<string, string[]>;
-  config: Config;
+  config: ValidatedConfig;
   createTimeSpan(msg: string, debug?: boolean): LoggerTimeSpan;
   data: any;
   debug: (msg: string) => void;
@@ -320,11 +321,13 @@ export interface CompilerBuildStats {
   rollupResults: RollupResults;
   sourceGraph?: BuildSourceGraph;
   componentGraph: BuildResultsComponentGraph;
-  collections: {
-    name: string;
-    source: string;
-    tags: string[];
-  }[];
+  collections: CompilerBuildStatCollection[];
+}
+
+export interface CompilerBuildStatCollection {
+  name: string;
+  source: string;
+  tags: string[][];
 }
 
 export interface CompilerBuildStatBundle {
@@ -424,10 +427,10 @@ export interface Cache {
 }
 
 export interface CollectionCompilerMeta {
-  collectionName?: string;
+  collectionName: string;
   moduleId?: string;
-  moduleDir?: string;
-  moduleFiles?: Module[];
+  moduleDir: string;
+  moduleFiles: Module[];
   global?: Module;
   compiler?: CollectionCompilerVersion;
   isInitialized?: boolean;
@@ -461,21 +464,6 @@ export interface CollectionBundleManifest {
 export interface CollectionDependencyManifest {
   name: string;
   tags: string[];
-}
-
-/** OLD WAY */
-export interface Collection {
-  collectionName?: string;
-  moduleDir?: string;
-  moduleFiles?: any[];
-  global?: any;
-  compiler?: CollectionCompiler;
-  isInitialized?: boolean;
-  hasExports?: boolean;
-  dependencies?: string[];
-  bundles?: {
-    components: string[];
-  }[];
 }
 
 export interface CollectionCompiler {